+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Remove dependency of `vtex.product-details` app.
+- Remove overpast `vtex-page-padding` class.
+
+### Changed
+- Update the `vtex.product-summary` and `vtex.store-components` versions.
 
 ## [1.0.1] - 2018-11-05
 ### Added